+++ chat.h
@@ -1,0 +1,5 @@
+#ifndef CHAT_H_
+#define CHAT_H_
+void * send_msg();
+void * receive_msg(void *msg);
+#endif
