+++ constants.h
@@ -1,0 +1,9 @@
+#ifndef CONSTANTS_H_
+#define CONSTANTS_H_
+
+#include <mqueue.h>
+#include <sys/stat.h>
+
+mqd_t q_send, q_receive;
+
+#endif
